+++ packages/typescript/src/messaging/client.ts
@@ -15,6 +15,7 @@ import { MessagesResource } from "./messages.js";
 import { ObservationPoliciesResource } from "./observation-policies.js";
 import { OperationsResource } from "./operations.js";
 import { ProfileResource } from "./profile.js";
+import { PrivacyResource } from "./privacy.js";
 import { QuickRepliesResource } from "./quick-replies.js";
 import { SessionsResource } from "./sessions.js";
 import { TemplatesResource } from "./templates.js";
@@ -32,6 +33,7 @@ export class MessagingClient {
   readonly messages: MessagesResource;
   readonly operations: OperationsResource;
   readonly profile: ProfileResource;
+  readonly privacy: PrivacyResource;
   readonly quickReplies: QuickRepliesResource;
   readonly templates: TemplatesResource;
   readonly webhooks: WebhooksResource;
@@ -61,6 +63,7 @@ export class MessagingClient {
     this.messages = new MessagesResource(transport);
     this.operations = new OperationsResource(transport);
     this.profile = new ProfileResource(transport);
+    this.privacy = new PrivacyResource(transport);
     this.quickReplies = new QuickRepliesResource(transport);
     this.templates = new TemplatesResource(transport);
     this.webhooks = new WebhooksResource(transport);